backend/requirements.txt
# Backend-Abhängigkeiten für Münz-Analyse API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
tensorflow>=2.13.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
scipy>=1.11.0
requests>=2.31.0
scikit-image>=0.21.0
matplotlib>=3.7.0
